vehicle diversion added; VrpData.setTime() removed;
git-svn-id: file:///srv/local/data/cleanRepo/trunk@27444 a0517364-e873-4167-b1c8-416daa5a6b91

+++ contribs/dvrp/src/main/java/org/matsim/contrib/dynagent/DynAgent.java
@@ -233,7 +233,7 @@ public class DynAgent
     public void notifyMoveOverNode(Id newLinkId)
     {
         double now = internalInterface.getMobsim().getSimTimer().getTimeOfDay();
-        dynLeg.movedOverNode(currentLinkId, newLinkId, (int)now);
+        dynLeg.movedOverNode(currentLinkId, newLinkId, now);
         currentLinkId = newLinkId;
     }
