+++ contribs/dvrp/src/main/java/org/matsim/contrib/dynagent/DynLeg.java
@@ -25,7 +25,7 @@ import org.matsim.api.core.v01.Id;
 public interface DynLeg
     extends DynAction
 {
-    void movedOverNode(Id oldLinkId, Id newLinkId, int time);
+    void movedOverNode(Id oldLinkId, Id newLinkId, double time);
 
 
     Id getCurrentLinkId();
